requirements.txt
Django==2.1.3
djangorestframework==3.7.7
psycopg2-binary==2.7.6.1
